Update calculate_age to use 'period' instead of 'duration', for clock-time based calculation
Add force_dates parameter to calculate_age to force defaults to align inputs to a date. There are some unexpected outputs for 'age' by year if a mixture of date and date times are used. As such, this behaviour is no longer a default and it attempts to convert to a date first. If this is intended, the date forcing can be forced off and a warning provided to keep the user aware.