Allow for header line checksfor iotools method for chunked reads
This avoids the first row reading the column names and causing oddities. This is only really a problem when chunking, since we dont want to skip on each chunk, we just want to skip the first time. However, we still need to apply this check since they will be split across cores and we do not drop the rows beforehand; the file is accessed right during the chunk.apply() step and that does not have a global skip parameter.